Enhanced database statistics with temporal analysis tracking
- Added earliest_analyzed_post and latest_analyzed_post fields to show actual Truth Social post timestamps for analyzed content
- Statistics now distinguish between all posts vs. analyzed posts date ranges
- Fixed date calculations to use actual post timestamps instead of database metadata
- Updated shitvault stats command to provide more informative temporal insights
- Enhanced documentation with examples of new statistics fields
- Provides visibility into analysis coverage and temporal gaps in processing
