+++ shit/db/__init__.py
@@ -1,0 +1,19 @@
+"""
+Database Shared Utilities
+Provides shared database operations for the Shitpost-Alpha project.
+"""
+
+from .database_client import DatabaseClient
+from .database_config import DatabaseConfig
+from .database_operations import DatabaseOperations
+from .database_utils import DatabaseUtils
+from .data_models import Base, model_to_dict
+
+__all__ = [
+    'DatabaseClient',
+    'DatabaseConfig', 
+    'DatabaseOperations',
+    'DatabaseUtils',
+    'Base',
+    'model_to_dict'
+]
