--- shit/tests/test_workflow_validation.py
@@ -1,188 +1,0 @@
-#!/usr/bin/env python3
-"""
-Workflow Validation Test
-Validate the new separated workflow architecture without requiring LLM API keys.
-"""
-
-import asyncio
-import pytest
-from datetime import datetime
-from shitvault.shitpost_db import ShitpostDatabase
-from shit.config.shitpost_settings import Settings
-
-@pytest.mark.asyncio
-async def test_workflow_architecture():
-    """Test the new workflow architecture."""
-    print("🏗️ Testing New Workflow Architecture")
-    print("=" * 50)
-    
-    db_manager = ShitpostDatabase()
-    settings = Settings()
-    
-    try:
-        await db_manager.initialize()
-        
-        print("1. ✅ Database initialization successful")
-        
-        # Test unprocessed posts query
-        print("\n2. Testing unprocessed posts query...")
-        posts = await db_manager.get_unprocessed_posts(
-            launch_date=settings.SYSTEM_LAUNCH_DATE,
-            limit=5
-        )
-        
-        print(f"   Found {len(posts)} unprocessed posts")
-        
-        if posts:
-            print("   Sample posts:")
-            for i, post in enumerate(posts[:2]):
-                print(f"     {i+1}. {post['text'][:50]}...")
-                print(f"        Engagement: {post.get('favourites_count', 0)} favs")
-                print(f"        Timestamp: {post['timestamp']}")
-        
-        # Test prediction deduplication
-        print("\n3. Testing prediction deduplication...")
-        if posts:
-            post_id = posts[0]['id']
-            exists = await db_manager.check_prediction_exists(post_id)
-            print(f"   Prediction exists for post {post_id}: {exists}")
-        
-        # Test with non-existent post
-        exists = await db_manager.check_prediction_exists(99999)
-        print(f"   Prediction exists for post 99999: {exists}")
-        
-        print("\n4. Workflow Architecture Benefits:")
-        print("   ✅ Ingestion and analysis are decoupled")
-        print("   ✅ Can run ingestion independently")
-        print("   ✅ Can run analysis independently")
-        print("   ✅ Launch date filtering prevents processing old posts")
-        print("   ✅ Prediction deduplication prevents duplicate analysis")
-        print("   ✅ Enhanced prediction model with Truth Social data")
-        print("   ✅ Rich engagement metrics for better analysis")
-        
-        print(f"\n🎉 Workflow architecture validation completed successfully!")
-        
-    except Exception as e:
-        print(f"❌ Test failed: {e}")
-        import traceback
-        traceback.print_exc()
-        raise
-    finally:
-        await db_manager.cleanup()
-
-@pytest.mark.asyncio
-async def test_launch_date_filtering():
-    """Test launch date filtering logic."""
-    print("\n📅 Testing Launch Date Filtering")
-    print("=" * 50)
-    
-    db_manager = ShitpostDatabase()
-    
-    try:
-        await db_manager.initialize()
-        
-        # Test different launch dates
-        test_cases = [
-            ("2025-01-01T00:00:00Z", "Before our posts - should find posts"),
-            ("2025-09-01T00:00:00Z", "After our posts - should find posts"),
-            ("2025-12-01T00:00:00Z", "Future date - should find no posts")
-        ]
-        
-        for launch_date, description in test_cases:
-            print(f"\n   Testing: {description}")
-            posts = await db_manager.get_unprocessed_posts(
-                launch_date=launch_date,
-                limit=5
-            )
-            print(f"   Found {len(posts)} posts")
-        
-        print(f"\n✅ Launch date filtering test completed!")
-        
-    except Exception as e:
-        print(f"❌ Test failed: {e}")
-        import traceback
-        traceback.print_exc()
-        raise
-    finally:
-        await db_manager.cleanup()
-
-@pytest.mark.asyncio
-async def test_enhanced_prediction_model():
-    """Test the enhanced prediction model structure."""
-    print("\n📊 Testing Enhanced Prediction Model")
-    print("=" * 50)
-    
-    db_manager = ShitpostDatabase()
-    
-    try:
-        await db_manager.initialize()
-        
-        # Test storing enhanced analysis (without LLM)
-        sample_post = {
-            'id': 1,
-            'replies_count': 2124,
-            'reblogs_count': 1697,
-            'favourites_count': 9496,
-            'upvotes_count': 9496,
-            'account_followers_count': 10564743,
-            'has_media': False,
-            'mentions': [],
-            'tags': [],
-            'text': 'GOOD NIGHT!!!'
-        }
-        
-        sample_analysis = {
-            'confidence': 0.8,
-            'assets': ['SPY', 'QQQ'],
-            'market_impact': {'SPY': 'bullish', 'QQQ': 'neutral'},
-            'thesis': 'Market sentiment analysis',
-            'llm_provider': 'test',
-            'llm_model': 'test-model',
-            'analysis_timestamp': datetime.utcnow()
-        }
-        
-        # Test enhanced analysis storage
-        prediction_id = await db_manager.store_analysis(
-            post_id="1",
-            analysis_data=sample_analysis,
-            post_data=sample_post
-        )
-        
-        if prediction_id:
-            print(f"   ✅ Successfully stored enhanced prediction: {prediction_id}")
-            
-            # Verify prediction exists
-            exists = await db_manager.check_prediction_exists(1)
-            print(f"   ✅ Prediction deduplication working: {exists}")
-        else:
-            print("   ⚠️  Prediction storage failed (might already exist)")
-        
-        print(f"\n✅ Enhanced prediction model test completed!")
-        
-    except Exception as e:
-        print(f"❌ Test failed: {e}")
-        import traceback
-        traceback.print_exc()
-        raise
-    finally:
-        await db_manager.cleanup()
-
-if __name__ == "__main__":
-    async def run_workflow_tests():
-        """Run workflow validation tests."""
-        print("🧪 Running Workflow Architecture Validation")
-        print("=" * 60)
-        
-        await test_workflow_architecture()
-        await test_launch_date_filtering()
-        await test_enhanced_prediction_model()
-        
-        print("\n🎉 All workflow validation tests completed successfully!")
-        print("\n📋 Summary:")
-        print("   ✅ Separated ingestion and analysis workflow")
-        print("   ✅ Launch date filtering prevents processing old posts")
-        print("   ✅ Prediction deduplication prevents duplicate analysis")
-        print("   ✅ Enhanced prediction model with Truth Social data")
-        print("   ✅ Ready for production deployment")
-    
-    asyncio.run(run_workflow_tests())
